refactor(mavvrik): address amane-cloudwiz review comments
proxy_server.py:
- Import Orchestrator, Settings, Uploader with AS aliases (MavvrikOrchestrator,
  MavvrikSettings, MavvrikUploader) per review
- Remove _ prefix from local variables (uploader, orchestrator)

litellm_logging.py + custom_logger_registry.py:
- Import Logger from litellm.integrations.mavvrik (package __init__)
  instead of the submodule path in both logging blocks and the registry
